+++ TravBotSharp.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TTWarsCore", "TTWarsCore\TT
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TravianOfficialCore", "TravianOfficialCore\TravianOfficialCore.csproj", "{D663785D-6F7D-4F37-9B4B-0D0C1E8067EF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject", "TestProject\TestProject.csproj", "{2B07661F-591D-49BC-AF91-F1F51BE2F488}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,18 @@ Global
 		{D663785D-6F7D-4F37-9B4B-0D0C1E8067EF}.Release|x64.Build.0 = Release|Any CPU
 		{D663785D-6F7D-4F37-9B4B-0D0C1E8067EF}.Release|x86.ActiveCfg = Release|Any CPU
 		{D663785D-6F7D-4F37-9B4B-0D0C1E8067EF}.Release|x86.Build.0 = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|x64.Build.0 = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Debug|x86.Build.0 = Debug|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|x64.ActiveCfg = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|x64.Build.0 = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|x86.ActiveCfg = Release|Any CPU
+		{2B07661F-591D-49BC-AF91-F1F51BE2F488}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
